+++ guillotina/contrib/cache/strategy.py
@@ -8,11 +8,12 @@ from guillotina.db.cache.base import BaseCache
 from guillotina.db.interfaces import ITransaction
 from guillotina.db.interfaces import ITransactionCache
 from guillotina.exceptions import NoPubSubUtility
+from guillotina.exceptions import NoChannelConfigured
 from guillotina.interfaces import ICacheUtility
 from guillotina.profile import profilable
 
 
-logger = logging.getLogger('guillotina')
+logger = logging.getLogger("guillotina")
 
 _default_size = 1024
 _basic_types = (bytes, str, int, float)
@@ -91,43 +92,38 @@ class BasicCache(BaseCache):
         try:
             if invalidate:
                 # A commit worked so we want to invalidate
-                keys_to_invalidate = await self._extract_invalidation_keys([
-                    (self._transaction.modified, 'modified'),
-                    (self._transaction.added, 'added'),
-                    (self._transaction.deleted, 'deleted')
-                ])
+                keys_to_invalidate = await self._extract_invalidation_keys(
+                    [
+                        (self._transaction.modified, "modified"),
+                        (self._transaction.added, "added"),
+                        (self._transaction.deleted, "deleted"),
+                    ]
+                )
                 await self.delete_all(keys_to_invalidate)
 
             if len(self._keys_to_publish) > 0 and self._utility._subscriber is not None:
                 asyncio.ensure_future(self.synchronize())
         except Exception:
-            logger.warning('Error closing connection', exc_info=True)
+            logger.warning("Error closing connection", exc_info=True)
 
     @profilable
     async def synchronize(self):
-        '''
+        """
         publish cache changes on redis
-        '''
+        """
         if self._utility._subscriber is None:
             raise NoPubSubUtility()
+        if app_settings.get("cache", {}).get("updates_channel", None) is None:
+            raise NoChannelConfigured()
         push = {}
         for obj, pickled in self._stored_objects:
-            val = {
-                'state': pickled,
-                'zoid': obj.__uuid__,
-                'tid': obj.__serial__,
-                'id': obj.__name__
-            }
+            val = {"state": pickled, "zoid": obj.__uuid__, "tid": obj.__serial__, "id": obj.__name__}
             if obj.__of__:
-                ob_key = self.get_key(
-                    oid=obj.__of__, id=obj.__name__, variant='annotation')
-                await self.set(
-                    val, oid=obj.__of__, id=obj.__name__, variant='annotation')
+                ob_key = self.get_key(oid=obj.__of__, id=obj.__name__, variant="annotation")
+                await self.set(val, oid=obj.__of__, id=obj.__name__, variant="annotation")
             else:
-                ob_key = self.get_key(
-                    container=obj.__parent__, id=obj.__name__)
-                await self.set(
-                    val, container=obj.__parent__, id=obj.__name__)
+                ob_key = self.get_key(container=obj.__parent__, id=obj.__name__)
+                await self.set(val, container=obj.__parent__, id=obj.__name__)
 
             if ob_key in self._keys_to_publish:
                 self._keys_to_publish.remove(ob_key)
@@ -135,11 +131,7 @@ class BasicCache(BaseCache):
 
         self._utility.ignore_tid(self._transaction._tid)
         await self._utility._subscriber.publish(
-            app_settings['cache']['updates_channel'],
+            app_settings["cache"]["updates_channel"],
             self._transaction._tid,
-            {
-                'tid': self._transaction._tid,
-                'keys': self._keys_to_publish,
-                'push': push
-            }
+            {"tid": self._transaction._tid, "keys": self._keys_to_publish, "push": push},
         )