+++ guillotina/exceptions.py
@@ -7,7 +7,6 @@ from zope.interface.interfaces import ComponentLookupError  # noqa pylint: disab
 
 
 class NoPermissionToAdd(Exception):
-
     def __init__(self, container, content_type):
         super().__init__()
         self.container = container
@@ -15,8 +14,8 @@ class NoPermissionToAdd(Exception):
 
     def __repr__(self):
         return "Not permission to add {content_type} on {container}".format(
-            content_type=self.content_type,
-            container=self.container)
+            content_type=self.content_type, container=self.container
+        )
 
 
 class InvalidContentType(Exception):
@@ -25,12 +24,10 @@ class InvalidContentType(Exception):
         self.content_type = content_type
 
     def __repr__(self):
-        return "Invalid type: {content_type}".format(
-            content_type=self.content_type)
+        return "Invalid type: {content_type}".format(content_type=self.content_type)
 
 
 class NotAllowedContentType(Exception):
-
     def __init__(self, container, content_type):
         super().__init__()
         self.container = container
@@ -38,19 +35,17 @@ class NotAllowedContentType(Exception):
 
     def __repr__(self):
         return "Not allowed {content_type} on {container}".format(
-            content_type=self.content_type,
-            container=self.container)
+            content_type=self.content_type, container=self.container
+        )
 
 
 class ConflictIdOnContainer(Exception):
-
     def __init__(self, pg_exc):
         super().__init__()
         self.pg_exc = pg_exc
 
     def __repr__(self):
-        msg = (getattr(self.pg_exc, 'detail', None) or
-               getattr(self.pg_exc, 'message', None) or self.pg_exc)
+        msg = getattr(self.pg_exc, "detail", None) or getattr(self.pg_exc, "message", None) or self.pg_exc
         return f"Conflict ID {msg}"
 
 
@@ -59,7 +54,6 @@ class UnRetryableRequestError(Exception):
 
 
 class PreconditionFailed(Exception):
-
     def __init__(self, container, precondition):
         super().__init__()
         self.container = container
@@ -67,8 +61,8 @@ class PreconditionFailed(Exception):
 
     def __repr__(self):
         return "Precondition Failed {precondition} on {path}".format(
-            precondition=self.precondition,
-            path=self.container)
+            precondition=self.precondition, path=self.container
+        )
 
 
 class RequestNotFound(Exception):
@@ -82,15 +76,18 @@ class TransactionNotFound(Exception):
     Lookup for the current request for request aware task failed
     """
 
+
 class ContainerNotFound(Exception):
     """
     Lookup for the current container for request aware task failed
     """
 
+
 class DatabaseNotFound(Exception):
     """
     Lookup for the current db for request aware task failed
     """
+
 
 @implementer(IUnauthorized)
 class Unauthorized(Exception):
@@ -98,28 +95,28 @@ class Unauthorized(Exception):
 
 
 class ConflictError(Exception):
-
-    def __init__(self, msg='', oid=None, txn=None, old_serial=None, writer=None):
+    def __init__(self, msg="", oid=None, txn=None, old_serial=None, writer=None):
         super().__init__()
         if oid is not None:
             conflict_summary = self.get_conflict_summary(oid, txn, old_serial, writer)
-            msg = f'{msg}.\n{conflict_summary}'
+            msg = f"{msg}.\n{conflict_summary}"
         super().__init__(msg)
 
     def get_conflict_summary(self, oid, txn, old_serial, writer):
         from guillotina.utils import get_current_request
+
         try:
             req = get_current_request()
         except RequestNotFound:
             req = None
-        max_attempts = app_settings.get('conflict_retry_attempts', 3)
-        attempts = getattr(req, '_retry_attempt', 0)
-        return f'''Object ID: {oid}
+        max_attempts = app_settings.get("conflict_retry_attempts", 3)
+        attempts = getattr(req, "_retry_attempt", 0)
+        return f"""Object ID: {oid}
 TID: {txn._tid}
 Old Object TID: {old_serial}
 Belongs to: {writer.of}
 Parent ID: {writer.id}
-Retries: {attempts}/{max_attempts}'''
+Retries: {attempts}/{max_attempts}"""
 
 
 class TIDConflictError(ConflictError):
@@ -127,9 +124,9 @@ class TIDConflictError(ConflictError):
 
 
 class RestartCommit(Exception):
-    '''
+    """
     Commits requires restart
-    '''
+    """
 
 
 class ConfigurationError(Exception):
@@ -146,7 +143,6 @@ class ComponentConfigurationError(ValueError, ConfigurationError):
 
 
 class ConfigurationConflictError(ConfigurationError):
-
     def __init__(self, conflicts):
         super().__init__()
         self._conflicts = conflicts
@@ -156,9 +152,9 @@ class ConfigurationConflictError(ConfigurationError):
         items = self._conflicts.items()
         items.sort()
         for discriminator, infos in items:
-            r.append("  For: %s" % (discriminator, ))
+            r.append("  For: %s" % (discriminator,))
             for info in infos:
-                for line in str(info).rstrip().split('\n'):
+                for line in str(info).rstrip().split("\n"):
                     r.append("    " + line)
 
         return "\n".join(r)
@@ -182,25 +178,21 @@ class DeserializationError(Exception):
 
     def __init__(self, errors):
         super().__init__()
-        self.msg = self.message = 'Error deserializing content'
+        self.msg = self.message = "Error deserializing content"
         self.errors = errors
 
     def __str__(self):
-        return '{} ({})'.format(
-            self.msg,
-            ujson.dumps(self.json_data()))  # pylint: disable=I1101
+        return "{} ({})".format(self.msg, ujson.dumps(self.json_data()))  # pylint: disable=I1101
 
     def json_data(self):
         errors = []
         for error in self.errors:
             # need to clean raw exceptions out of this list here...
             error = error.copy()
-            if 'error' in error:
-                error.pop('error')
+            if "error" in error:
+                error.pop("error")
             errors.append(error)
-        return {
-            'deserialization_errors': errors
-        }
+        return {"deserialization_errors": errors}
 
 
 class ValueDeserializationError(Exception):
@@ -224,11 +216,18 @@ class FileNotFoundException(Exception):
 
 
 class ServerClosingException(Exception):
-    '''
+    """
     Server closing, can not perform action
-    '''
+    """
+
 
 class NoPubSubUtility(Exception):
-    '''
+    """
     No PubSub Utility found
-    '''
+    """
+
+
+class NoChannelConfigured(Exception):
+    """
+    No Channel Configured
+    """